app/requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
gunicorn==20.1.0
psycopg2-binary==2.9.1
django-extensions
pillow
sorl-thumbnail==12.8